optimised data types used

+++ v21_gfp_exact_sched_test/2processors/algorithm_move.cpp
@@ -10,12 +10,12 @@
 
 
 // check which jobs interfere lower priority jobs
-void update_interferance_flags(state& s, const uint_fast8_t m, const uint_fast8_t* perm) {
+void update_interferance_flags(state& s, const unsigned short m, const unsigned short* perm) {
     
     s.updatePendJobsNum();
     
     if (s.pendJobsNum > m)
-        for (uint_fast8_t i = 0; i < m; i++)
+        for (unsigned short i = 0; i < m; i++)
             s.interferred[perm[i]] = true;
 }
 
@@ -26,12 +26,12 @@ void update_interferance_flags(state& s, const uint_fast8_t m, const uint_fast8_
 // Deadline check
 bool deadline_miss(const state& s, const TS& ts, const bool verbose) {
     
-    const uint_fast8_t N = s.n;
+    const unsigned short N = s.n;
     
     if (s.d(N-1, ts.P[N-1], ts.D[N-1]) < s.c[N-1]) {
         if (verbose) {
             cout << "Failure state:" << endl;
-            for (uint_fast8_t i = 0; i < N; i++) {
+            for (unsigned short i = 0; i < N; i++) {
                 cout << "c[" << (int)i << "]: " << (int)s.c[i] << " d[" << (int)i << "]: " << (int)s.d(N-1, ts.P[N-1], ts.D[N-1]) << " p[" << (int)i << "]" << (int)s.p[i] << endl;
             }
         }
@@ -52,20 +52,20 @@ bool deadline_miss(const state& s, const TS& ts, const bool verbose) {
 // -1 - deadline miss
 // 0 - continue traversal of successors
 // 1 - schedule can be discarded
-uint_fast8_t algorithm_move(state& s, const TS& ts, const uint_fast8_t m, bool verbose) {
+short int algorithm_move(state& s, const TS& ts, const unsigned short m, bool verbose) {
 
-    uint_fast8_t N = ts.n;
-    uint_fast8_t* perm = new uint_fast8_t [ts.n];
+    unsigned short N = ts.n;
+    unsigned short* perm = new unsigned short [ts.n];
 
     sortTasksByPriorities(s, perm);
     update_interferance_flags(s, m, perm);
 
-    uint_fast8_t dt = deltaT(s, m, perm);
+    unsigned short dt = deltaT(s, m, perm);
     
     if (!condition_job_interference(s, m, dt, perm)) {
         // interf. cond. violated
-        for (uint_fast8_t i = 0; i < m; i++) s.c[perm[i]] = max(s.c[perm[i]] - dt, 0);
-        for (uint_fast8_t i = 0; i < N; i++) s.p[i] = max(s.p[i] - dt, 0); // ????????? to recheck
+        for (unsigned short i = 0; i < m; i++) s.c[perm[i]] = max(s.c[perm[i]] - dt, 0);
+        for (unsigned short i = 0; i < N; i++) s.p[i] = max(s.p[i] - dt, 0); // ????????? to recheck
         if (deadline_miss(s, ts, verbose)) {delete [] perm; return -1;}
         else {
             //cout << "pruned by condition_job_interference(s, m, dt, perm)" << endl;
@@ -74,25 +74,25 @@ uint_fast8_t algorithm_move(state& s, const TS& ts, const uint_fast8_t m, bool v
         }
     } else {
         // interf. cond. holds
-        const uint_fast8_t previousCk = s.c[N-1];
-        for (uint_fast8_t i = 0; i < N; i++) s.jobCanBeReleasedBefore[i] = ((s.p[i] == 0)?true:false);
+        const unsigned short previousCk = s.c[N-1];
+        for (unsigned short i = 0; i < N; i++) s.jobCanBeReleasedBefore[i] = ((s.p[i] == 0)?true:false);
         
         //conditionC_cri_tau_i(s, ts, m, perm); // set .releaseAtEarliest flags
         
-        for (uint_fast8_t i = 0; i < m; i++) s.c[perm[i]] = max(s.c[perm[i]] - dt, 0);
-        for (uint_fast8_t i = 0; i < N; i++) s.p[i] = max(s.p[i] - dt, 0);
-        for (uint_fast8_t i = 0; i < N; i++) s.timePassedFromLastRelease[i] = s.timePassedFromLastRelease[i] + dt;
+        for (unsigned short i = 0; i < m; i++) s.c[perm[i]] = max(s.c[perm[i]] - dt, 0);
+        for (unsigned short i = 0; i < N; i++) s.p[i] = max(s.p[i] - dt, 0);
+        for (unsigned short i = 0; i < N; i++) s.timePassedFromLastRelease[i] = s.timePassedFromLastRelease[i] + dt;
         
         //if (s.tau2_releaseNextJobBefore > 0) s.tau2_releaseNextJobBefore -= dt;
         //if (s.tau3_releaseNextJobBefore > 0) s.tau3_releaseNextJobBefore -= dt;
-        for (uint_fast8_t i = 0; i < ts.n; i++) {
+        for (unsigned short i = 0; i < ts.n; i++) {
             if (s.tauI_releaseNextJobBefore[i] > 0) s.tauI_releaseNextJobBefore[i] -= dt;
             // TO CHECK: should the computation of dt be changed? can be optimized?
         }
         
         if (deadline_miss(s, ts, verbose)) {delete [] perm; return -1;}
         
-        for (uint_fast8_t i = 0; i < N; i++) s.prevState_processorAvailableForTau_i[i] = s.processorAvailableForTau_i[i];
+        for (unsigned short i = 0; i < N; i++) s.prevState_processorAvailableForTau_i[i] = s.processorAvailableForTau_i[i];
         
         if ((previousCk > 0) && (s.c[N-1] == 0)) {
             //cout << "pruned: state is discarded, as analyzed job completes execution" << endl;
